chore(cli): unify derivation options (#19)
This PR:

1. Fixes #8 - the `DERIVE KEY` command will not be used in the CLI, and
all `EXPORT KEY` and `SIGN` commands default to a sane derivation path
and are configured to derive keys from the master seed on every
invocation.
2. In completing #8, incidentally this fixes #14 as the derivation path
is now configured when creating the signer and then has the Ethereum
address exported from the card to store in a local variable.